requirements.txt
langchain==0.3.25
openai==1.79.0
pgvector==0.3.6
psycopg[binary]==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
python-dotenv==0.21.0
python-docx==1.1.2
langchain-openai==0.3.17
langchain-core==0.3.60
langchain-postgres==0.0.14
langchain-text-splitters==0.3.8
langchain_community == 0.3.20
langgraph==0.4.5
langgraph-swarm==0.0.11
gradio==5.31.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.31.0